Secure File Sharing/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose
email-validator
aiofiles
itsdangerous